requirements.txt
streamlit
requests
python-dotenv
pillow
firebase-admin